Rendering multiple models with multiple textures
Added another texture to testing (background.jpg).

Camera now follows the one dynamic object in the scene.
Press C to switch to manual camera movement again. (Added to readme controls)

Engine constructor now loads all textures in rsc\images into a list.
Then initializes the entities we want to be in the scene.
initEntities() now loads the models from src\models one at a time for each entity we specify.

Can no longer cycle through models rendered as that was just for testing and we can now render any models we want.

Model class now has an id parameter to keep track of the model name associated to it's mesh.
Model class now has a position variable so the camera knows where to point.

Renderer no longer needs shouldRender bool as it now renders everything in the model list and we control what is added to the list explicitly in the Engine class.

Renderer::render now takes in a list of textures so we can associate each model with a different texture.

Added initial movement to the dynamic boxcar in Simulate to better show physX collisions.
Simulation step now updates the position of each dynamic model every simulation step.