onTrigger working and physx loops through list of vehicles
Increased the tile size of the arena and took out the sky box.

Simulator now takes into it's constructor the list of vehicles so that in initialization physx can create a pxVehicle for each vehicle we want to model.

actor->userData still holds const char* id.  Couldn't figure out how to extract the id or compare the models in the pose update function when setting userData to vehicle->carModel.