Joshua (#26)
* Experimenting with rendering multiple physx

Added teapot to scene as a PxSphereGeometry so the bounding box is a square.

Added "id" to each model so i can identify which physx actor matches which model when updating their transformations.

* Experimenting with rendering multiple physx actors

Added teapot to scene as a PxSphereGeometry so the bounding box is a square.

Added "id" to each model so i can identify which physx actor matches which model when updating their transformations.

* Rendering multiple dynamic actors with collisions

Moved the starting position of the camera to behind the player vehicle.

Removed the MoveType enum from model since it was no longer used.
Added an id parameter to model.  Id is used to match the physx collision model to the graphics model.

Added deltaSec to the stepPhysics function to match the change in time to our framerate.

Added an ai1 vehicle to test how to implement additional, separate vehicles.

* Fixed a actor flag bug and reset Vehicle class

* onTrigger working and physx loops through list of vehicles

Increased the tile size of the arena and took out the sky box.

Simulator now takes into it's constructor the list of vehicles so that in initialization physx can create a pxVehicle for each vehicle we want to model.

actor->userData still holds const char* id.  Couldn't figure out how to extract the id or compare the models in the pose update function when setting userData to vehicle->carModel.

* Driving controls in the Vehicle class

Vehicle class now has a VehicleController struct that gets passed to the simulator to process driving controls.  Player vehicle works as normal and the AI vehicle can now be moved separately by calling it's appropriate move functions.

Controller class directly calls the player vehicles movement functions instead of storing an array of inputs.

Reorganize a bit of the driving functions in Simulate.  Renamed incrementDrivingMode to setDriveMode to make more sense with our project.

Vehicle 'forward' parameter not updating currently.  It's only used to determine start direction.