Resize and moved the vehicle model
Vehicle now sits very close (but not right on) the tile plane.

Changed the phsyX vehicle model to have the bottom of the chassis sit half-way up the wheel instead of the car body sit on top of the wheel.  Helps with stability and avoids flipping when turning.

Vehicle graphics model is scaled down and translated down (hard coded in Vehcile::setModelMatrix) to match the physX vehicle models.