Revised driving specs and vehicle size (#46)
Tuned the driving aspects to better fit out vision. Got rid of the floatiness.

Shrunk the vehicle size and increased the tile size to get a better contrast of scale.

Pulled the camera in tighter to decrease FOV.  Still not on a spring system.

More testing needs to be done to ensure new driving parameters do not flip the vehicle over (resulting in the vehicle falling through the map).