Add both graphic and physX pickup models to the scene
PickupManager sets up the locations for every pickup per arena.  Adds those locations to a queue of toBeAddedPickups then past to the simulator to construct a physx body as well as adding the model to the renderer.

Collision detection with pickups calls the pickup managers handlePickupOnCollision with a location where the collision occurred.  The pickup currently does not disappear  upon collision.