Power ups (#61)
* Revised driving specs and vehicle size

Tuned the driving aspects to better fit out vision. Got rid of the floatiness.

Shrunk the vehicle size and increased the tile size to get a better contrast of scale.

Pulled the camera in tighter to decrease FOV.  Still not on a spring system.

More testing needs to be done to ensure new driving parameters do not flip the vehicle over (resulting in the vehicle falling through the map).

* Added Battery to scene to restore vehicle energy

Vehicle now has energy drain and restore functionality when driving over tiles.

Trigger collision with a box in the scene to restore battery.

* fixing errors

* Updated branch from master

* Add both graphic and physX pickup models to the scene

PickupManager sets up the locations for every pickup per arena.  Adds those locations to a queue of toBeAddedPickups then past to the simulator to construct a physx body as well as adding the model to the renderer.

Collision detection with pickups calls the pickup managers handlePickupOnCollision with a location where the collision occurred.  The pickup currently does not disappear  upon collision.

* Pulled back camera and decreased tile energy drain