Fixed: Crashes on pickup collision
Multiple pickup actors and renderables were being created. Collisions would remove the pickup but the duplicate actor would trigger another collision (trying to remove a pickup that was already removed).

Created some safeguards such as always supplying an actors->userData with a value besides the default nullptr so that any checks while iterating through the actors don't mistakenly match with 0 (null), as well as a ERROR console msg in the PickupManager::handlePickupOnCollision to notify when a collided pickup could not be found, returning a dummy pickup to the simulator to avoid instant crashes.