Vehicle Boost and Trap functionality (#75)
Vehicle now has boost functions.  Applying the boost takes a duration in frames per second.  Boost can be released and applied later if boost duration remains.
Vehicle is boosted to a hard-coded max speed that can be adjusted.

Vehicle now has trap functions that slow it down (as if in a slow trap).
Applies a strong force backwards on the vehicle to keep in under a minimum speed.
Min speed is hard-coded and can be adjusted.

For testing purposes you can use T to trigger boost and G to trigger trap.