Add barriers around arena and removed vehicle::straighten() (#94)
Added parameters to physx createDrivablePlane() to set the planes normal and distance.

Added 4 barriers to the arena using the createDrivablePlane() function.  Distance is gathered from the arena size / 2 times the arenaTileWidth so it should match for any size arena.

Removed experimental function Vehicle::straighten().  Tried to counter-balance hard turns but turning the other way for 1 frame but it ended up making the AI vehicles wobble on straight runs.