Camera rework (#104)
* Camera Boom Arm created and Joystick control testing

Camera sits on a boom and the boom arm tries to catch up to the player when driving.

Added joystick controls but only the X (on playstation 4 controller) accelerates the car.  Every joystick button should output which number button they are on the console.

* Turned initial volume to 50%

* Camera Swing and Lots of DevUI parameters

Camera now swings to the back of the vehicle.

Major camera parameters added to the DevUI.

Some Handling parameters added to the DevUI.

Added sections, borders, and transparent background to DevUI.

Fixed the barriers not being created on the inside of the border walls.